# apps/delivery-api/docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: CLEVER Route Server Delivery API
  version: 0.1.0
  license:
    name: UNLICENSED
    identifier: LicenseRef-UNLICENSED
  summary: CLEVER route server delivery API HTTP contract
  description: >-
    Machine-readable contract for CLEVER route server communication with current
    admin clients, legacy Shopify compatibility routes, native driver mobile
    clients, and operational health checks. Human behavior notes live beside
    this file in docs/api/*.md.
  x-contract-revision: 2026-05-14
  x-source-repository: EVNSolution/clever-route-server
servers:
  - url: http://localhost:3000
    description: Local development server
  - url: https://{deliveryServerHost}
    description: Deployed delivery server
    variables:
      deliveryServerHost:
        default: delivery.example.com
security: []
...
paths:
...
  /admin/route-plans:
    get:
      tags: [Shopify Embedded Admin]
      summary: List route plans for the authenticated shop
      operationId: listAdminRoutePlans
      x-source-file: src/routes/admin-route-plans.routes.ts
      security:
        - ShopifySessionToken: []
      responses:
        '200':
          description: Route plans for the authenticated shop.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoutePlansListEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [Shopify Embedded Admin]
      summary: Create a draft route plan from selected Shopify orders
      operationId: createAdminRoutePlan
      x-source-file: src/routes/admin-route-plans.routes.ts
      security:
        - ShopifySessionToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRoutePlanRequest'
      responses:
        '201':
          description: Draft route plan created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoutePlanEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
  /admin/route-plans/{routePlanId}:
    parameters:
      - $ref: '#/components/parameters/RoutePlanId'
    get:
      tags: [Shopify Embedded Admin]
      summary: Read a route plan detail for the authenticated shop
      operationId: getAdminRoutePlan
      x-source-file: src/routes/admin-route-plans.routes.ts
      security:
        - ShopifySessionToken: []
      responses:
        '200':
          description: Route plan detail, stops, and optional geometry.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoutePlanDetailEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Shopify Embedded Admin]
      summary: Delete a route plan for the authenticated shop
      operationId: deleteAdminRoutePlan
      x-source-file: src/routes/admin-route-plans.routes.ts
      security:
        - ShopifySessionToken: []
      responses:
        '200':
          description: Route plan deletion result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericDataEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /admin/route-plans/{routePlanId}/stops:
    parameters:
      - $ref: '#/components/parameters/RoutePlanId'
    patch:
      tags: [Shopify Embedded Admin]
      summary: Replace ordered stops for a route plan
      operationId: updateAdminRoutePlanStops
      x-source-file: src/routes/admin-route-plans.routes.ts
      security:
        - ShopifySessionToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRoutePlanStopsRequest'
      responses:
        '200':
          description: Updated route plan detail.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoutePlanDetailEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
...
components:
  securitySchemes:
    ShopifySessionToken:
      type: http
      scheme: bearer
      bearerFormat: Shopify App Bridge session JWT
      description: Server-verified Shopify embedded app session token.
    DriverAccessToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Server-issued short-lived driver JWT with audience clever-delivery-driver.
    ShopifyWebhookHmac:
      type: apiKey
      in: header
      name: X-Shopify-Hmac-Sha256
      description: Base64 HMAC generated by Shopify over the raw request body.
  parameters:
    RoutePlanId:
      name: routePlanId
      in: path
      required: true
      schema:
        type: string
...
  responses:
    BadRequest:
      description: Invalid request payload, query, path, or headers.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Unauthorized:
      description: Missing or invalid bearer token or HMAC.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
    NotFound:
      description: Requested resource was not found under the authenticated boundary.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Conflict:
      description: Request conflicts with existing route/order assignment state.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
      additionalProperties: true
    ErrorEnvelope:
      type: object
      required: [data, error]
      properties:
        data:
          type: 'null'
        error:
          $ref: '#/components/schemas/Error'
    GenericDataEnvelope:
      type: object
      required: [data, error]
      properties:
        data:
          type: object
          additionalProperties: true
        error:
          type: 'null'
...
    CreateRoutePlanRequest:
      type: object
      required: [name, planDate, depot, orders]
      properties:
        name:
          type: string
        planDate:
          format: date
          type: string
        depot:
          type: object
          required: [address, latitude, longitude]
          properties:
            address:
              type:
                - string
                - 'null'
            latitude:
              type:
                - number
                - 'null'
            longitude:
              type:
                - number
                - 'null'
        routeScope:
          $ref: '#/components/schemas/RouteScope'
        orders:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/RoutePlanOrderInput'
      additionalProperties: false
    RouteScope:
      type: object
      required: [deliveryDate, deliverySession, routeScopeKey, serviceType]
      properties:
        deliveryDate:
          format: date
          type: string
        deliverySession:
          enum: [DAY, EVENING, PICKUP]
          type: string
        routeScopeKey:
          type: string
        serviceType:
          $ref: '#/components/schemas/DeliveryServiceType'
        timeWindowStart:
          type:
            - string
            - 'null'
          pattern: '^\d{2}:\d{2}$'
        timeWindowEnd:
          type:
            - string
            - 'null'
          pattern: '^\d{2}:\d{2}$'
    RoutePlanOrderInput:
      type: object
      required: [name, shippingAddress, shopifyOrderGid]
      properties:
        shopifyOrderGid:
          type: string
        name:
          type: string
        email:
          type:
            - string
            - 'null'
        phone:
          type:
            - string
            - 'null'
        financialStatus:
          type:
            - string
            - 'null'
        fulfillmentStatus:
          type:
            - string
            - 'null'
        processedAt:
          format: date-time
          type:
            - string
            - 'null'
        totalPriceAmount:
          type:
            - string
            - 'null'
        currencyCode:
          type:
            - string
            - 'null'
        recipientName:
          type:
            - string
            - 'null'
        shippingAddress:
          $ref: '#/components/schemas/ShippingAddress'
        latitude:
          type:
            - number
            - 'null'
        longitude:
          type:
            - number
            - 'null'
        deliveryArea:
          type:
            - string
            - 'null'
        deliveryDay:
          type:
            - string
            - 'null'
        attributes:
          type: array
          items:
            $ref: '#/components/schemas/KeyValueAttribute'
        rawPayload:
          type: object
          additionalProperties: true
        routeScopeKey:
          type:
            - string
            - 'null'
        deliveryDate:
          format: date
          type:
            - string
            - 'null'
        deliverySession:
          type:
            - string
            - 'null'
          enum: [DAY, EVENING, PICKUP, null]
        serviceType:
          oneOf:
            - $ref: '#/components/schemas/DeliveryServiceType'
            - type: 'null'
        timeWindowStart:
          type:
            - string
            - 'null'
        timeWindowEnd:
          type:
            - string
            - 'null'
      additionalProperties: true
    UpdateRoutePlanStopsRequest:
      type: object
      required: [stops]
      properties:
        stops:
          type: array
          items:
            type: object
            required: [sequence, shopifyOrderGid]
            properties:
              deliveryStopId:
                type:
                  - string
                  - 'null'
              sequence:
                type: integer
                minimum: 1
              shopifyOrderGid:
                type: string
      additionalProperties: false
    RoutePlanSummary:
      type: object
      required: [id, name, status]
      properties:
        id:
          type: string
        name:
          type: string
        status:
          type: string
        planDate:
          format: date
          type:
            - string
            - 'null'
        stopsCount:
          type: integer
        missingCoordinates:
          type: integer
        deliveryAreas:
          type: array
          items:
            type: string
        deliveryDays:
          type: array
          items:
            type: string
        depot:
          type: object
          additionalProperties: true
        createdAt:
          format: date-time
          type: string
        updatedAt:
          format: date-time
          type: string
      additionalProperties: true
    RoutePlanEnvelope:
      type: object
      required: [data, error]
      properties:
        data:
          type: object
          required: [routePlan]
          properties:
            routePlan:
              $ref: '#/components/schemas/RoutePlanSummary'
        error:
          type: 'null'
    RoutePlansListEnvelope:
      type: object
      required: [data, error]
      properties:
        data:
          type: object
          required: [routePlans]
          properties:
            routePlans:
              type: array
              items:
                $ref: '#/components/schemas/RoutePlanSummary'
        error:
          type: 'null'
    RoutePlanDetailEnvelope:
      type: object
      required: [data, error]
      properties:
        data:
          type: object
          required: [routePlan, stops]
          properties:
            routePlan:
              $ref: '#/components/schemas/RoutePlanSummary'
            routeGeometry:
              type:
                - object
                - 'null'
              additionalProperties: true
            routeStopPoints:
              type: array
              items:
                type: object
                additionalProperties: true
            stops:
              type: array
              items:
                type: object
                additionalProperties: true
          additionalProperties: true
        error:
          type: 'null'
...
    ShippingAddress:
      type: object
      properties:
        address1:
          type:
            - string
            - 'null'
        address2:
          type:
            - string
            - 'null'
        city:
          type:
            - string
            - 'null'
        province:
          type:
            - string
            - 'null'
        postalCode:
          type:
            - string
            - 'null'
        countryCode:
          type:
            - string
            - 'null'
      additionalProperties: false
    KeyValueAttribute:
      type: object
      required: [key, value]
      properties:
        key:
          type: string
        value:
          type: string
      additionalProperties: false
    DeliveryServiceType:
      enum: [DELIVERY, EVENING_DELIVERY, PICKUP]
      type: string